feat: add invalidParameter error
